Clean up the structure
