+++ web/src/table_tennis/components/GameModeSwitch.js
@@ -1,6 +1,10 @@
 import { Box, Switch } from "@chakra-ui/react";
+import { useGameDispatch, useGameState } from "../GameContext";
 
-const GameModeSwitch = ({ isDouble, dispatch }) => {
+const GameModeSwitch = () => {
+  const dispatch = useGameDispatch();
+  const state = useGameState();
+
   const handleToggle = (event) => {
     dispatch({
       type: "game_mode_toggled",
@@ -12,7 +16,7 @@ const GameModeSwitch = ({ isDouble, dispatch }) => {
     <Box py="4">
       Single&nbsp;
       <Switch
-        isChecked={isDouble}
+        isChecked={state.isDouble}
         onChange={handleToggle}
         size="lg"
         id="game-mode"
