+++ web/src/table_tennis/components/index.js
@@ -3,3 +3,5 @@ export { default as ScoreSelector } from "./ScoreSelector";
 export { default as PlayerSelector } from "./PlayerSelector";
 export { default as LoadingPlaceholder } from "./LoadingPlaceholder";
 export { default as DoubleDownButton } from "./DoubleDownButton";
+export { default as SubmitButton } from "./SubmitButton";
+export { default as AddPlayerButton } from "./AddPlayerButton";
