+++ web/src/table_tennis/index.js
@@ -1,0 +1,10 @@
+import { ContextProvider } from "./GameContext";
+import Screen from "./Screen";
+
+const TableTennis = () => (
+  <ContextProvider>
+    <Screen />
+  </ContextProvider>
+);
+
+export default TableTennis;
